Moved controller switching routine to the interface. algorithm_blocks is now totally ros-free. Confirmed also that _utils is actually used.
